+++ packages/iov-keycontrol/src/hdpaths.ts
@@ -68,12 +68,23 @@ export class HdPaths {
   }
 
   /**
-   * The default MetaMask derivation path
+   * The default Ethereum derivation path
+   *
+   * This is compatible to MetaMask and Trezor.
    *
    * What MetaMask calls the account is a BIP44 address index: m/44'/60'/0'/0/<account>
    * (see https://github.com/MetaMask/eth-hd-keyring/blob/018a11a3a2/index.js#L8)
    */
-  public static metamaskHdKeyTree(account: number): ReadonlyArray<Slip10RawIndex> {
+  public static ethereum(account: number): ReadonlyArray<Slip10RawIndex> {
     return HdPaths.bip44(60, 0, 0, account);
+  }
+
+  /**
+   * The default MetaMask derivation path
+   *
+   * @deprecated use HdPaths.ethereum
+   */
+  public static metamaskHdKeyTree(account: number): ReadonlyArray<Slip10RawIndex> {
+    return HdPaths.ethereum(account);
   }
 }